e1 & e3 finished (with complete tests and everything commented)

+++ DS-61-08-A2/src/e1/Ghost.java
@@ -5,6 +5,7 @@ public class Ghost extends Resident {
     House Hou;
 
     public Ghost(String name, String surname, String house, int horcruxes) {
+        //pass every value through the constructor
         Name = name;
         Surname = surname;
         Horcruxes = horcruxes;
@@ -12,12 +13,14 @@ public class Ghost extends Resident {
         category = Category.Ghost;
     }
 
+    //gets the category string for the print functions
     @Override
     String getCategory() { return category.toString() + " of " + Hou.toString() + " "; }
 
+    //gets the reward double fo the print functions
     @Override
     double getReward() {
-        int r = Horcruxes * 80;
+        double r = Horcruxes * 80;
 
         if (Hou == House.Slytherin) r *= 2;
