+++ DS-61-08-A2/src/e1/Student.java
@@ -5,6 +5,7 @@ public class Student extends Resident {
     House Hou;
 
     public Student(String name, String surname, String house, int horcruxes) {
+        //pass every value through the constructor
         Name = name;
         Surname = surname;
         Horcruxes = horcruxes;
@@ -12,12 +13,14 @@ public class Student extends Resident {
         category = Category.Student;
     }
 
+    //gets the category string for the print functions
     @Override
     String getCategory() { return category.toString() + " of " + Hou.toString() + " "; }
 
+    //gets the reward double fo the print functions
     @Override
     double getReward() {
-        int r = Horcruxes * 90;
+        double r = Horcruxes * 90;
 
         if (Hou == House.Slytherin) r *= 2;
