+++ DS-61-08-A2/src/e1/Teacher.java
@@ -5,6 +5,7 @@ public class Teacher extends Staff {
     Subject Sub;
 
     public Teacher(String name, String surname, String subject, int horcruxes) {
+        //pass every value through the constructor
         Name = name;
         Surname = surname;
         Horcruxes = horcruxes;
@@ -12,14 +13,16 @@ public class Teacher extends Staff {
         category = Category.Teacher;
     }
 
+    //gets the category string for the print functions
     @Override
     String getCategory() {
         return category.toString() + " of " + Sub.toString() + " ";
     }
 
+    //gets the reward double fo the print functions
     @Override
     double getReward() {
-        int r = Horcruxes * 50;
+        double r = Horcruxes * 50;
 
         if (Sub == Subject.Defence)  r *= 0.75;
